requirements.txt

google-generativeai
faiss-cpu
python-docx
python-pptx
pdfminer.six
pytesseract
Pillow
openai-whisper
yt-dlp
streamlit
tqdm
pytest
